resolver: move ValidateTargetURI from grpcutil to avoid new leaf-package deps
